+++ src/main/java/cc/cassian/raspberry/mixin/CoconutSliceItemMixin.java
@@ -1,0 +1,66 @@
+/*
+Based off the Milk Bottle item from Farmer's Delight. Its license is available below.
+MIT License
+
+Copyright (c) 2020 vectorwing
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+ */
+package cc.cassian.raspberry.mixin;
+
+import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.entity.living.MobEffectEvent;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+import samebutdifferent.ecologics.item.CoconutSliceItem;
+
+import java.util.ArrayList;
+import java.util.Iterator;
+
+@Pseudo
+@Mixin(CoconutSliceItem.class)
+public class CoconutSliceItemMixin {
+    @Redirect(method = "finishUsingItem", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/player/Player;removeAllEffects()Z"))
+    private boolean mixin(Player consumer) {
+        Iterator<MobEffectInstance> itr = consumer.getActiveEffects().iterator();
+        ArrayList<MobEffect> compatibleEffects = new ArrayList<>();
+
+        while(itr.hasNext()) {
+            MobEffectInstance effect = itr.next();
+            if (effect.isCurativeItem(new ItemStack(Items.MILK_BUCKET))) {
+                compatibleEffects.add(effect.getEffect());
+            }
+        }
+
+        if (!compatibleEffects.isEmpty()) {
+            MobEffectInstance selectedEffect = consumer.getEffect(compatibleEffects.get(consumer.level.random.nextInt(compatibleEffects.size())));
+            if (selectedEffect != null && !MinecraftForge.EVENT_BUS.post(new MobEffectEvent.Remove(consumer, selectedEffect))) {
+                consumer.removeEffect(selectedEffect.getEffect());
+            }
+        }
+        return false;
+    }
+}